pyproject.toml
[project]
name = "style2fab"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yamato Miyatake", email = "ymt.mytk@gmail.com" }
]
dependencies = [
    "django",
    "djangorestframework",
    "scipy",
    "matplotlib",
    "requests",
    "pandas",
    "scikit-learn",
    "torch==2.3.1+cu121",
    "torchvision==0.18.1+cu121",
    "torchaudio==2.3.1+cu121",
    "kaolin==0.17.0",
    "tqdm",
    "pymeshlab",
    "open-clip-torch>=2.32.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

# PyTorch CUDA 12.1 用の index リポジトリ
[[tool.rye.sources]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
type = "index"

# Kaolin 用の find-links リポジトリ
[[tool.rye.sources]]
name = "kaolin-py"
url = "https://nvidia-kaolin.s3.us-east-2.amazonaws.com/torch-2.3.1_cu121.html"
type = "find-links"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/style2fab"]
